move config parsing for bulk case up to RulesUtils
